pyproject.toml
[project]
name = "mmfiles"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.1",
    "bitsandbytes>=0.47.0",
    "colpali-engine>=0.3.12",
    "gradio>=5.47.1",
    "librosa>=0.11.0",
    "mcp>=1.14.1",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "soundfile>=0.13.1",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
]

[project.scripts]
mmfiles-mcp = "mmfiles:main_mcp"
mmfiles = "mmfiles:main_app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
torch = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true
